final refactor before merging
